+++ CHANGES.md
@@ -101,10 +101,16 @@ Authorization:
 -   Added per-record authorization around the `/records/<recordid>/history` endpoint
 -   Added per-record authorization around the `/records/<recordid>/history/<eventId>` endpoint
 
-Search
+Registry:
+
+-   Made events record the correct user id
+-   Fixed Registry History API Performance Issue when limit=1 & Updated Registry History API Document
+
+Search:
 
 -   Began work on a NodeJS-based search API to replace the scala one
 -   Allow fetch region record by region id
+-   Use a list of acronyms for publishers that have a 'Department of' in their name
 
 Minions:
 
@@ -113,7 +119,6 @@ Minions:
 Others:
 
 -   Use a "Year" column from a CSV file to extract a temporal extent
--   Fixed Registry History API Performance Issue when limit=1 & Updated Registry History API Document
 -   Added tentative documentation/scripting for building and running on microk8s.
 -   Added tentative documentation/scripting for building and running on k3d.